requirements.txt
# requirements

PyQt5==5.9.2
pyserial
fbs